pyproject.toml
[tool.poetry]
name = "trapper-discord-bot"
version = "1.0.0"
description = "A scalable Discord bot framework"
authors = ["Your Name"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^2.3.2"
python-dotenv = "^1.0.0"
aiohttp = "^3.9.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
